Separate comment section from the analysis page
