+++ services/agora/src/components/post/comments/group/CommentGroup.vue
@@ -15,20 +15,9 @@
       class="commentListFlex"
     >
       <ZKCard
-        v-if="mode === 'analysis' && aiSummary !== undefined"
-        padding="1rem"
-        class="commentItemBackground"
-      >
-        <CommentConsensusSummary
-          :summary="aiSummary"
-          :selected-cluster-key="selectedClusterKey"
-        />
-      </ZKCard>
-
-      <ZKCard
         v-for="commentItem in commentItemList"
         :id="commentItem.opinionSlugId"
-        :key="commentItem.opinionSlugId + '-' + selectedClusterKey"
+        :key="commentItem.opinionSlugId"
         padding="0rem"
         class="commentItemBackground"
         :class="{
@@ -37,8 +26,6 @@
         }"
       >
         <CommentItem
-          :mode="mode"
-          :selected-cluster-key="selectedClusterKey"
           :comment-item="commentItem"
           :post-slug-id="postSlugId"
           :comment-slug-id-liked-map="commentSlugIdLikedMap"
@@ -58,10 +45,9 @@
 </template>
 
 <script setup lang="ts">
-import type { OpinionItem, PolisKey, VotingAction } from "src/shared/types/zod";
+import type { OpinionItem, VotingAction } from "src/shared/types/zod";
 import CommentItem from "./item/CommentItem.vue";
 import ZKCard from "src/components/ui-library/ZKCard.vue";
-import CommentConsensusSummary from "../../analysis/CommentConsensusSummary.vue";
 
 const emit = defineEmits(["deleted", "mutedComment", "changeVote"]);
 
@@ -70,11 +56,8 @@ function changeVote(vote: VotingAction, opinionSlugId: string) {
 }
 
 defineProps<{
-  mode: "comment" | "analysis";
-  selectedClusterKey: PolisKey | undefined;
   commentItemList: OpinionItem[];
   postSlugId: string;
-  aiSummary?: string;
   initialCommentSlugId: string;
   commentSlugIdLikedMap: Map<string, "agree" | "disagree">;
   isPostLocked: boolean;
